requirements.txt
aioquic==1.2.0
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
dateparser==1.2.0
et_xmlfile==2.0.0
Flask==3.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
json-stream==2.3.2
json-stream-rs-tokenizer==0.4.26
kaitaistruct==0.10
ldap3==2.9.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mitmproxy==11.0.2
mitmproxy-windows==0.10.7
mitmproxy2swagger==0.14.0
mitmproxy_rs==0.10.7
msgpack==1.1.0
numpy==2.2.0
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
psutil==6.1.1
publicsuffix2==2.20191221
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydivert==2.1.0
Pygments==2.18.0
pylsqpack==0.3.18
pyOpenSSL==24.3.0
pyparsing==3.2.0
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
ruff==0.8.6
selenium==4.27.1
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tornado==6.4.2
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
urwid==2.6.16
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zstandard==0.23.0
